+++ db/migrate/001_add_autowatch_involved_issue_to_user.rb
@@ -1,0 +1,5 @@
+class AddAutowatchInvolvedIssueToUser < ActiveRecord::Migration
+  def change
+    add_column :user_preferences, :autowatch_involved_issue, :boolean, default: true, null: false
+  end
+end