+++ test/unit/watcher_test.rb
@@ -11,6 +11,32 @@ class WatcherTest < ActiveSupport::TestCase
     @author = User.find(1)
     @assigned_user = User.find(2)
     @changing_user = User.find(4)
+
+    Setting.plugin_additionals = ActionController::Parameters.new(
+      issue_autowatch_involved: 1
+    )
+  end
+
+  def test_new_issue_with_no_autowatch
+    Setting.plugin_additionals = ActionController::Parameters.new(
+      issue_autowatch_involved: 0
+    )
+
+    User.current = @author
+    issue = Issue.generate(author_id: @author.id)
+    issue.save
+    assert_equal 0, issue.watchers.count
+    assert !issue.watched_by?(@author)
+  end
+
+  def test_new_issue_with_no_autowatch_by_user
+    User.current = @author
+    User.current.pref.update_attribute :autowatch_involved_issue, false
+
+    issue = Issue.generate(author_id: @author.id)
+    issue.save
+    assert_equal 0, issue.watchers.count
+    assert !issue.watched_by?(@author)
   end
 
   def test_new_issue_with_author_watch_only